send notification
This is a notification function that is triggered when an order has arrived from an end user, a notification is send to a respective device using a unique device ID that is captured when ever a user logins in to his device or changes to a another device. The notification carries a payload that includes the sender, product, type, quantity and the app icon.